Adapt execution
