--- libs/execution/src/lib/constraints/executors/regex-constraint-executor.ts
@@ -1,30 +1,0 @@
-// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
-//
-// SPDX-License-Identifier: AGPL-3.0-only
-
-import { type InternalValueRepresentation } from '@jvalue/jayvee-language-server';
-
-import { type ExecutionContext } from '../../execution-context';
-import { implementsStatic } from '../../util/implements-static-decorator';
-import { type ConstraintExecutor } from '../constraint-executor';
-import { type TypedConstraintExecutorClass } from '../typed-constraint-executor-class';
-
-@implementsStatic<TypedConstraintExecutorClass>()
-export class RegexConstraintExecutor implements ConstraintExecutor {
-  public static readonly type = 'RegexConstraint';
-
-  isValid(
-    value: InternalValueRepresentation,
-    context: ExecutionContext,
-  ): boolean {
-    if (typeof value !== 'string') {
-      return false;
-    }
-
-    const regex = context.getPropertyValue(
-      'regex',
-      context.valueTypeProvider.Primitives.Regex,
-    );
-    return regex.test(value);
-  }
-}
